requirements.txt
accelerate==1.12.0
aider-install==0.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
blivet==3.12.1
Brlapi==0.8.7
cbor2==5.8.0
certifi==2025.7.9
cffi==2.0.0
cfgv==3.5.0
chardet==7.0.1
charset-normalizer==3.4.3
click==8.3.1
cockpit @ file:///builddir/build/BUILD/cockpit-357-build/cockpit-357/tmp/wheel/cockpit-357-py3-none-any.whl#sha256=474f6d2f07f5e1fc6d7a1714574c3620ecb9769fd645f220ab4b2314f02dc195
-e git+ssh://git@github.com/Joeavaib/llm.git@f8f3e95bdc790a9fc4208d450bb53687fd7ac7f4#egg=context_engine&subdirectory=context_engine
-e git+https://github.com/Joeavaib/partner.git@0751b2371ccce5d7514bce32144d168c18064489#egg=contextmachine&subdirectory=meta-orchestrator
crypt_r==3.13.1
cryptography==46.0.5
cssselect==1.3.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cupshelpers==1.0
dasbus==1.7
datasets==4.5.0
dbus-python==1.4.0
dbus_next==0.2.3
dill==0.4.0
distlib==0.4.0
distro==1.9.0
dnf==4.24.0
docker==7.1.0
docstring_parser==0.17.0
evalplus==0.3.1
faiss-cpu==1.13.2
fastapi==0.128.3
fastcore==1.12.23
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.25.0
fire==0.7.1
frozenlist==1.8.0
fsspec==2025.10.0
gguf==0.17.1
ghapi==1.0.13
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.3.7
human-eval==1.0.3
humanize==4.13.0
hyperframe==6.1.0
identify==2.6.17
idna==3.10
iniconfig==2.3.0
inkex==1.4.0
iso639==0.1.4
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
louis==3.33.0
lxml==6.0.1
-e git+https://github.com/Joeavaib/maestro.git@daf121609bcdd80336966851de5b526bbf7cb75d#egg=maestro
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modal==1.3.5
mpmath==1.3.0
multidict==6.7.1
multipledispatch==1.0.0
multiprocess==0.70.18
networkx==3.6.1
nftables==0.1
nodeenv==1.10.0
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
olefile==0.47
ollama==0.6.1
openai==2.20.0
packaging==25.0
pandas==3.0.0
peft==0.18.1
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
pipx==1.8.0
platformdirs==4.9.4
pluggy==1.6.0
ply==3.11
portalocker==3.2.0
pre_commit==4.5.1
productmd==1.50
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pwquality==1.4.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
PyMuPDF==1.26.7
pyparsing==3.1.2
pyparted==3.13.0
pyperclip==1.11.0
pyserial==3.5
PySocks==1.7.1
pytest==9.0.2
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-dotenv==1.2.2
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.22
python-pam==2.0.2
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.13-build/kernel-6.18.13/linux-6.18.13-200.fc43.x86_64/tools/net/ynl
qdrant-client==1.16.2
rank-bm25==0.2.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rich==14.3.2
rpds-py==0.27.0
rpm==6.0.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scour==0.38.2
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sentence-transformers==5.2.3
sentencepiece==0.2.1
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
shellingham==1.5.4
simpleline==1.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sos==4.11.0
soupsieve==2.8
starlette==0.50.0
stop-sequencer==1.2.3
swebench==4.1.0
sympy==1.14.0
synchronicity==0.11.1
systemd-python==235
tempdir==0.7.1
tenacity==9.1.4
termcolor==3.3.0
threadpoolctl==3.6.0
tinycss2==1.5.1
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
tree-sitter==0.25.2
tree-sitter-python==0.25.0
triton==3.6.0
trl==0.28.0
typer==0.24.1
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
unidiff==0.7.5
uritemplate==4.2.0
urllib3==2.6.3
userpath==1.9.2
uv==0.9.29
uvicorn==0.40.0
virtualenv==21.1.0
watchfiles==1.1.1
webencodings==0.5.1
wget==3.2
xkbregistry==0.3
xxhash==3.6.0
yarl==1.22.0
